backend/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
beautifulsoup4==4.12.3
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.0
dotenv==0.9.9
ecdsa==0.19.0
fastapi==0.115.6
google==3.0.0
google-ai-generativelanguage==0.6.10
google-api-core==2.24.0
google-api-python-client==2.158.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
grpcio==1.69.0
grpcio-status==1.69.0
h11==0.14.0
httplib2==0.22.0
httptools==0.6.4
idna==3.10
jose==1.0.0
packaging==24.2
pillow==11.1.0
proto-plus==1.25.0
protobuf==5.29.3
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
pyparsing==3.2.1
pytesseract==0.3.13
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.41.3
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==14.1
